Moved Monitor and DB classes to the same protocol file
